Slightly refactor routes into urls.tsx
